move clip overflow
