+++ fury/tests/test_utils.py
@@ -8,8 +8,8 @@ from fury.utils import (map_coordinates_3d_4d,
                         rotate, vtk, vertices_from_actor,
                         compute_bounds, set_input,
                         update_actor, get_actor_from_primitive,
-                        get_bounds, clip_overflow)
-from fury import actor, window, utils, ui
+                        get_bounds)
+from fury import actor, window, utils
 import fury.primitive as fp
 
 
@@ -450,19 +450,3 @@ def test_get_bounds():
     actor.SetMapper(mapper)
     compute_bounds(actor)
     npt.assert_equal(get_bounds(actor), test_bounds)
-
-
-def test_clip_overflow():
-    text = ui.TextBlock2D(text="", position=(50, 50), color=(1, 0, 0))
-    rectangle = ui.Rectangle2D(position=(50, 50), size=(100, 50))
-
-    sm = window.ShowManager()
-    sm.scene.add(rectangle, text)
-
-    text.message = "Hello"
-    clip_overflow(text, rectangle.size[0])
-    npt.assert_equal("Hello", text.message)
-
-    text.message = "Hello wassup"
-    clip_overflow(text, rectangle.size[0])
-    npt.assert_equal("Hello was...", text.message)
