+++ fury/ui.py
@@ -10,11 +10,50 @@ import abc
 from fury.data import read_viz_icons
 from fury.interactor import CustomInteractorStyle
 from fury.io import load_image
-from fury.utils import set_input, rotate, clip_overflow
+from fury.utils import set_input, rotate
 from fury.actor import grid
 
 
 TWO_PI = 2 * np.pi
+
+
+def clip_overflow(textblock, width):
+    """Clips overflowing text of TextBlock2D with respect to width.
+
+    Parameters
+    ----------
+    textblock : TextBlock2D
+        The textblock object whose text needs to be clipped.
+    width : int
+        Required width of the clipped text.
+
+    Returns
+    -------
+    clipped text : str
+        Clipped version of the text.
+    """
+    original_str = textblock.message
+    start_ptr = 0
+    end_ptr = len(original_str)
+    prev_bg = textblock.have_bg
+    textblock.have_bg = False
+
+    if textblock.size[0] == width or textblock.size[0] <= width:
+        textblock.have_bg = prev_bg
+        return original_str
+
+    while start_ptr < end_ptr:
+        mid_ptr = (start_ptr + end_ptr)//2
+        textblock.message = original_str[:mid_ptr] + "..."
+        if textblock.size[0] < width:
+            start_ptr = mid_ptr
+        elif textblock.size[0] > width:
+            end_ptr = mid_ptr
+
+        if mid_ptr == (start_ptr + end_ptr)//2 or\
+           textblock.size[0] == width:
+            textblock.have_bg = prev_bg
+            return textblock.message
 
 
 class UI(object, metaclass=abc.ABCMeta):
@@ -1224,7 +1263,7 @@ class TextBlock2D(UI):
             bg_size = self.background.size
             if bb_size[0] > bg_size[0] or bb_size[1] > bg_size[1]:
                 warn("Font size exceeds background bounding box."
-                " Font Size will not be updated.", RuntimeWarning)
+                     " Font Size will not be updated.", RuntimeWarning)
                 self.actor.SetTextScaleModeToProp()
                 self.actor.SetPosition2(*bg_size)
 
@@ -1477,8 +1516,8 @@ class TextBlock2D(UI):
                 return size
             else:
                 warn("TextBlock2D must be added to the scene before "
-                "querying its size while TextScaleMode is set to None.", 
-                RuntimeWarning)
+                     "querying its size while TextScaleMode is set to None.",
+                     RuntimeWarning)
 
         return self.actor.GetPosition2()
