+++ fury/utils.py
@@ -1073,43 +1073,3 @@ def get_bounds(actor):
 
     """
     return actor.GetMapper().GetInput().GetBounds()
-
-
-def clip_overflow(textblock, width):
-    """Clips overflowing text of TextBlock2D with respect to width.
-
-    Parameters
-    ----------
-    textblock : TextBlock2D
-        The textblock object whose text needs to be clipped.
-    width : int
-        Required width of the clipped text.
-
-    Returns
-    -------
-    clipped text : str
-        Clipped version of the text.
-    """
-    original_str = textblock.message
-    start_ptr = 0
-    end_ptr = len(original_str)
-    prev_bg = textblock.have_bg
-    textblock.have_bg = False
-
-    if textblock.size[0] == width or textblock.size[0] <= width:
-        textblock.have_bg = prev_bg
-        return original_str
-
-    else:
-        while start_ptr < end_ptr:
-            mid_ptr = (start_ptr + end_ptr)//2
-            textblock.message = original_str[:mid_ptr] + "..."
-            if textblock.size[0] < width:
-                start_ptr = mid_ptr
-            elif textblock.size[0] > width:
-                end_ptr = mid_ptr
-
-            if mid_ptr == (start_ptr + end_ptr)//2 or\
-               textblock.size[0] == width:
-                textblock.have_bg = prev_bg
-                return textblock.message
